requirements_production.txt
# Core LangGraph dependencies
langgraph>=0.2.50
langgraph-sdk>=0.1.36
langchain>=0.3.13
langchain-core>=0.3.25
langchain-openai>=0.2.10
langchain-community>=0.3.13

# API Framework
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
pydantic>=2.10.4
pydantic-settings>=2.6.1

# HTTP Client
httpx>=0.28.1

# Database (if using Supabase)
supabase>=2.11.0

# Environment management
python-dotenv>=1.0.1

# Monitoring (optional but recommended)
langsmith>=0.2.6

# Utilities
aiofiles>=24.1.0
python-multipart>=0.0.20